+++ Changelog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Autodetect PyPy executables in [#617](https://github.com/PyO3/maturin/pull/617)
+* auditwheel: add `libz.so.1` to whitelisted libraries in [#625](https://github.com/PyO3/maturin/pull/625)
+* auditwheel: detect musl libc in [#629](https://github.com/PyO3/maturin/pull/629)
 
 ## [0.11.3] - 2021-08-25